src/requirements.txt
matplotlib==3.8.4
numpy==2.2.1
pandas==2.2.3
Pillow==11.0.0
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.4.2
timm==1.0.12
torch==2.5.1+cu124
torchvision==0.20.1+cu124
umap_learn==0.5.7
